Refactor to abstract trait methods instead of instanceof check
